requirements.txt
# Core BigQuery and AI
google-cloud-bigquery>=3.11.0
bigframes>=0.20.0
google-cloud-documentai>=2.20.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Visualization and Dashboard
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=9.5.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0